+++ src/col_to_rgb.c
@@ -148,7 +148,7 @@ char *col_name[] = {
 //  apply(mat, 1, \(x) paste0("{", paste(x, collapse = ", "), "},")) |>
 //  cat(sep = "\n")
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-static int col_int[][4] = {
+static uint8_t col_int[][4] = {
   {255, 255, 255,   0}, // NA
   {255, 255, 255,   0}, // transparent
   {255, 255, 255, 255}, // white
@@ -818,6 +818,59 @@ static int col_int[][4] = {
 #define hex2nibble(x) ( (((x) & 0xf) + ((x) >> 6) + ((x >> 6) << 3)) & 0xf )
 
 
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// Core C function
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+void col_to_rgb(const char *col, uint8_t ptr[4]) {
+  if (col[0] == '#') {
+    switch(strlen(col)) {
+    case 9: 
+      ptr[0] = (uint8_t)( (hex2nibble(col[1]) << 4) + hex2nibble(col[2]) );
+      ptr[1] = (uint8_t)( (hex2nibble(col[3]) << 4) + hex2nibble(col[4]) );
+      ptr[2] = (uint8_t)( (hex2nibble(col[5]) << 4) + hex2nibble(col[6]) );
+      ptr[3] = (uint8_t)( (hex2nibble(col[7]) << 4) + hex2nibble(col[8]) );
+      break;
+    case 7: 
+      ptr[0] = (uint8_t)( (hex2nibble(col[1]) << 4) + hex2nibble(col[2]) );
+      ptr[1] = (uint8_t)( (hex2nibble(col[3]) << 4) + hex2nibble(col[4]) );
+      ptr[2] = (uint8_t)( (hex2nibble(col[5]) << 4) + hex2nibble(col[6]) );
+      ptr[3] = 255;
+      break;
+    case 5: 
+      ptr[0] = hex2nibble(col[1]) * (16 + 1);
+      ptr[1] = hex2nibble(col[2]) * (16 + 1);
+      ptr[2] = hex2nibble(col[3]) * (16 + 1);
+      ptr[3] = hex2nibble(col[4]) * (16 + 1);
+      break;
+    case 4: 
+      ptr[0] = hex2nibble(col[1]) * (16 + 1);
+      ptr[1] = hex2nibble(col[2]) * (16 + 1);
+      ptr[2] = hex2nibble(col[3]) * (16 + 1);
+      ptr[3] = 255;
+      break;
+    default:
+      Rf_error("col_to_rgb_(): Hex notation error: %s", col);
+    }
+  } else {
+    
+    int idx = hash_color((const unsigned char *)col);
+    // don't need to do a full string comparison, as the probability of
+    // an incorrect color name (e.g. 'bluexx') hashing to a color
+    // that starts with 'blu' seems incredibly remote.  
+    // Probably worth testing though to figure out what is needed to 
+    // actually cause a collision here. (and how much of the strings should
+    // be compared to detect it)
+    if (idx < 0 || idx > 658 || memcmp(col, col_name[idx], 2) != 0) {
+      Rf_error("col_to_rgb_(): Not a valid color name: %s", col);
+    }
+    memcpy(ptr, col_int[idx], 4 * sizeof(uint8_t));
+  }
+}
+
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// R interface
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 SEXP col_to_rgb_(SEXP cols_) {
   
   int nprotect = 0;
@@ -865,53 +918,16 @@ SEXP col_to_rgb_(SEXP cols_) {
   // String colors
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   for(int i = 0; i < n; i++) {
-    const char *str = CHAR(STRING_ELT(cols_, i));
+    const char *col = CHAR(STRING_ELT(cols_, i));
+    uint8_t tmp[4];
+    col_to_rgb(col, tmp);
     
-    if (str[0] == '#') {
-      switch(strlen(str)) {
-      case 9: 
-        ptr[0] = (hex2nibble(str[1]) << 4) + hex2nibble(str[2]);
-        ptr[1] = (hex2nibble(str[3]) << 4) + hex2nibble(str[4]);
-        ptr[2] = (hex2nibble(str[5]) << 4) + hex2nibble(str[6]);
-        ptr[3] = (hex2nibble(str[7]) << 4) + hex2nibble(str[8]);
-        break;
-      case 7: 
-        ptr[0] = (hex2nibble(str[1]) << 4) + hex2nibble(str[2]);
-        ptr[1] = (hex2nibble(str[3]) << 4) + hex2nibble(str[4]);
-        ptr[2] = (hex2nibble(str[5]) << 4) + hex2nibble(str[6]);
-        ptr[3] = 255;
-        break;
-      case 5: 
-        ptr[0] = hex2nibble(str[1]) * (16 + 1);
-        ptr[1] = hex2nibble(str[2]) * (16 + 1);
-        ptr[2] = hex2nibble(str[3]) * (16 + 1);
-        ptr[3] = hex2nibble(str[4]) * (16 + 1);
-        break;
-      case 4: 
-        ptr[0] = hex2nibble(str[1]) * (16 + 1);
-        ptr[1] = hex2nibble(str[2]) * (16 + 1);
-        ptr[2] = hex2nibble(str[3]) * (16 + 1);
-        ptr[3] = 255;
-        break;
-      default:
-        Rf_error("col_to_rgb_(): Hex notation error: %s", str);
-      }
-      ptr += 4;
-    } else {
-      
-      int idx = hash_color((const unsigned char *)str);
-      // don't need to do a full string comparison, as the probability of
-      // an incorrect color name (e.g. 'bluexx') hashing to a color
-      // that starts with 'blu' seems incredibly remote.  
-      // Probably worth testing though to figure out what is needed to 
-      // actually cause a collision here. (and how much of the strings should
-      // be compared to detect it)
-      if (idx < 0 || idx > 658 || memcmp(str, col_name[idx], 2) != 0) {
-        Rf_error("col_to_rgb_(): Not a valid color name: %s", str);
-      }
-      memcpy(ptr, col_int[idx], 4 * sizeof(int));
-      ptr += 4;
-    }
+    ptr[0] = (int)tmp[0]; 
+    ptr[1] = (int)tmp[1]; 
+    ptr[2] = (int)tmp[2]; 
+    ptr[3] = (int)tmp[3]; 
+    
+    ptr += 4;
   }