+++ NEWS.md
@@ -1,3 +1,12 @@
+
+# colorfast 1.0.1
+
+* Exporting core C functions for access by other packages with `LinkingTo`
+* Export macros for accessing individual color components: 
+  `CF_RED()`, `CF_GREEN()`, `CF_BLUE()`, `CF_ALPHA()`
+* Add `int_to_col()` for converting integer value to hex color
+* Add `set_alpha()`
+
 # colorfast 1.0.0
 
 * Initial release of 'colorfast' (previously known as 'colourlookup')